[policies] Switch to common method for determing OS
This uses /etc/os-release exclusively.
It also requires the module names to match the OS id or id_like
exactly as written in os-release

Drop Atomic host as there is none in RHEL8

Resolves: #2113
