+++ sos/policies/redhat.py
@@ -17,8 +17,6 @@ from sos.policies import LinuxPolicy, PackageManager, PresetDefaults
 from sos import _sos as _
 from sos.options import SoSOptions
 
-OS_RELEASE = "/etc/os-release"
-
 
 class RedHatPolicy(LinuxPolicy):
     distro = "Red Hat"
@@ -80,18 +78,6 @@ class RedHatPolicy(LinuxPolicy):
         self.PATH += os.pathsep + "/usr/local/sbin"
         self.set_exec_path()
         self.load_presets()
-
-    @classmethod
-    def check(cls, remote=''):
-        """This method checks to see if we are running on Red Hat. It must be
-        overriden by concrete subclasses to return True when running on a
-        Fedora, RHEL or other Red Hat distribution or False otherwise.
-
-        If `remote` is provided, it should be the contents of a remote host's
-        os-release, or comparable, file to be used in place of the locally
-        available one.
-        """
-        return False
 
     def check_usrmove(self, pkgs):
         """Test whether the running system implements UsrMove.
@@ -255,309 +241,5 @@ No changes will be made to system configuration.
 RH_API_HOST = "https://access.redhat.com"
 RH_FTP_HOST = "ftp://dropbox.redhat.com"
 
-
-class RHELPolicy(RedHatPolicy):
-    distro = RHEL_RELEASE_STR
-    vendor = "Red Hat"
-    vendor_url = "https://access.redhat.com/support/"
-    msg = _("""\
-This command will collect diagnostic and configuration \
-information from this %(distro)s system and installed \
-applications.
-
-An archive containing the collected information will be \
-generated in %(tmpdir)s and may be provided to a %(vendor)s \
-support representative.
-""" + disclaimer_text + "%(vendor_text)s\n")
-    _upload_url = RH_FTP_HOST
-    _upload_user = 'anonymous'
-    _upload_directory = '/incoming'
-
-    def __init__(self, sysroot=None, init=None, probe_runtime=True,
-                 remote_exec=None):
-        super(RHELPolicy, self).__init__(sysroot=sysroot, init=init,
-                                         probe_runtime=probe_runtime,
-                                         remote_exec=remote_exec)
-        self.register_presets(rhel_presets)
-
-    @classmethod
-    def check(cls, remote=''):
-        """Test to see if the running host is a RHEL installation.
-
-            Checks for the presence of the "Red Hat Enterprise Linux"
-            release string at the beginning of the NAME field in the
-            `/etc/os-release` file and returns ``True`` if it is
-            found, and ``False`` otherwise.
-
-            :returns: ``True`` if the host is running RHEL or ``False``
-                      otherwise.
-        """
-
-        if remote:
-            return cls.distro in remote
-
-        if not os.path.exists(OS_RELEASE):
-            return False
-
-        with open(OS_RELEASE, "r") as f:
-            for line in f:
-                if line.startswith("NAME"):
-                    (name, value) = line.split("=")
-                    value = value.strip("\"'")
-                    if value.startswith(cls.distro):
-                        return True
-        return False
-
-    def prompt_for_upload_user(self):
-        if self.commons['cmdlineopts'].upload_user:
-            return
-        # Not using the default, so don't call this prompt for RHCP
-        if self.commons['cmdlineopts'].upload_url:
-            super(RHELPolicy, self).prompt_for_upload_user()
-            return
-        if self.case_id:
-            self.upload_user = input(_(
-                "Enter your Red Hat Customer Portal username (empty to use "
-                "public dropbox): ")
-            )
-
-    def get_upload_url(self):
-        if self.commons['cmdlineopts'].upload_url:
-            return self.commons['cmdlineopts'].upload_url
-        if (not self.case_id or not self.upload_user or not
-                self.upload_password):
-            # Cannot use the RHCP. Use anonymous dropbox
-            self.upload_user = self._upload_user
-            self.upload_directory = self._upload_directory
-            self.upload_password = None
-            return RH_FTP_HOST
-        else:
-            rh_case_api = "/hydra/rest/cases/%s/attachments"
-            return RH_API_HOST + rh_case_api % self.case_id
-
-    def _get_upload_headers(self):
-        if self.get_upload_url().startswith(RH_API_HOST):
-            return {'isPrivate': 'false', 'cache-control': 'no-cache'}
-        return {}
-
-    def get_upload_url_string(self):
-        if self.get_upload_url().startswith(RH_API_HOST):
-            return "Red Hat Customer Portal"
-        return self.upload_url or RH_FTP_HOST
-
-    def get_upload_user(self):
-        # if this is anything other than dropbox, annonymous won't work
-        if self.upload_url != RH_FTP_HOST:
-            return self.upload_user
-        return self._upload_user
-
-    def dist_version(self):
-        try:
-            rr = self.package_manager.all_pkgs_by_name_regex("redhat-release*")
-            pkgname = self.pkgs[rr[0]]["version"]
-            if pkgname[0] == "4":
-                return 4
-            elif pkgname[0] in ["5Server", "5Client"]:
-                return 5
-            elif pkgname[0] == "6":
-                return 6
-            elif pkgname[0] == "7":
-                return 7
-            elif pkgname[0] == "8":
-                return 8
-        except Exception:
-            pass
-        return False
-
-    def probe_preset(self):
-        # Emergency or rescue mode?
-        for target in ["rescue", "emergency"]:
-            if self.init_system.is_running("%s.target" % target):
-                return self.find_preset(CB)
-        # Package based checks
-        if self.pkg_by_name("satellite-common") is not None:
-            return self.find_preset(RH_SATELLITE)
-        if self.pkg_by_name("rhosp-release") is not None:
-            return self.find_preset(RHOSP)
-        if self.pkg_by_name("cfme") is not None:
-            return self.find_preset(RH_CFME)
-        if self.pkg_by_name("ovirt-engine") is not None or \
-                self.pkg_by_name("vdsm") is not None:
-            return self.find_preset(RHV)
-
-        # Vanilla RHEL is default
-        return self.find_preset(RHEL)
-
-
-class CentOsPolicy(RHELPolicy):
-    distro = "CentOS"
-    vendor = "CentOS"
-    vendor_url = "https://www.centos.org/"
-
-
-ATOMIC = "atomic"
-ATOMIC_RELEASE_STR = "Atomic"
-ATOMIC_DESC = "Red Hat Enterprise Linux Atomic Host"
-
-atomic_presets = {
-    ATOMIC: PresetDefaults(name=ATOMIC, desc=ATOMIC_DESC, note=NOTE_TIME,
-                           opts=_opts_verify)
-}
-
-
-class RedHatAtomicPolicy(RHELPolicy):
-    distro = "Red Hat Atomic Host"
-    msg = _("""\
-This command will collect diagnostic and configuration \
-information from this %(distro)s system.
-
-An archive containing the collected information will be \
-generated in %(tmpdir)s and may be provided to a %(vendor)s \
-support representative.
-""" + disclaimer_text + "%(vendor_text)s\n")
-
-    containerzed = True
-    container_runtime = 'docker'
-    container_image = 'registry.access.redhat.com/rhel7/support-tools'
-    sos_path_strip = '/host'
-    container_version_command = 'rpm -q sos'
-
-    def __init__(self, sysroot=None, init=None, probe_runtime=True,
-                 remote_exec=None):
-        super(RedHatAtomicPolicy, self).__init__(sysroot=sysroot, init=init,
-                                                 probe_runtime=probe_runtime,
-                                                 remote_exec=remote_exec)
-        self.register_presets(atomic_presets)
-
-    @classmethod
-    def check(cls, remote=''):
-
-        if remote:
-            return cls.distro in remote
-
-        atomic = False
-        if ENV_HOST_SYSROOT not in os.environ:
-            return atomic
-        host_release = os.environ[ENV_HOST_SYSROOT] + cls._redhat_release
-        if not os.path.exists(host_release):
-            return False
-        try:
-            for line in open(host_release, "r").read().splitlines():
-                atomic |= ATOMIC_RELEASE_STR in line
-        except IOError:
-            pass
-        return atomic
-
-    def probe_preset(self):
-        if self.pkg_by_name('atomic-openshift'):
-            return self.find_preset(RHOCP)
-
-        return self.find_preset(ATOMIC)
-
-    def create_sos_container(self):
-        _cmd = ("{runtime} run -di --name {name} --privileged --ipc=host"
-                " --net=host --pid=host -e HOST=/host -e NAME={name} -e "
-                "IMAGE={image} -v /run:/run -v /var/log:/var/log -v "
-                "/etc/machine-id:/etc/machine-id -v "
-                "/etc/localtime:/etc/localtime -v /:/host {image}")
-        return _cmd.format(runtime=self.container_runtime,
-                           name=self.sos_container_name,
-                           image=self.container_image)
-
-    def set_cleanup_cmd(self):
-        return 'docker rm --force sos-collector-tmp'
-
-
-class RedHatCoreOSPolicy(RHELPolicy):
-    distro = "Red Hat CoreOS"
-    msg = _("""\
-This command will collect diagnostic and configuration \
-information from this %(distro)s system.
-
-An archive containing the collected information will be \
-generated in %(tmpdir)s and may be provided to a %(vendor)s \
-support representative.
-""" + disclaimer_text + "%(vendor_text)s\n")
-
-    containerized = True
-    container_runtime = 'podman'
-    container_image = 'registry.redhat.io/rhel8/support-tools'
-    sos_path_strip = '/host'
-    container_version_command = 'rpm -q sos'
-
-    def __init__(self, sysroot=None, init=None, probe_runtime=True,
-                 remote_exec=None):
-        super(RedHatCoreOSPolicy, self).__init__(sysroot=sysroot, init=init,
-                                                 probe_runtime=probe_runtime,
-                                                 remote_exec=remote_exec)
-
-    @classmethod
-    def check(cls, remote=''):
-
-        if remote:
-            return 'CoreOS' in remote
-
-        coreos = False
-        if ENV_HOST_SYSROOT not in os.environ:
-            return coreos
-        host_release = os.environ[ENV_HOST_SYSROOT] + cls._redhat_release
-        try:
-            for line in open(host_release, 'r').read().splitlines():
-                coreos |= 'Red Hat Enterprise Linux CoreOS' in line
-        except IOError:
-            pass
-        return coreos
-
-    def probe_preset(self):
-        # As of the creation of this policy, RHCOS is only available for
-        # RH OCP environments.
-        return self.find_preset(RHOCP)
-
-    def create_sos_container(self):
-        _cmd = ("{runtime} run -di --name {name} --privileged --ipc=host"
-                " --net=host --pid=host -e HOST=/host -e NAME={name} -e "
-                "IMAGE={image} -v /run:/run -v /var/log:/var/log -v "
-                "/etc/machine-id:/etc/machine-id -v "
-                "/etc/localtime:/etc/localtime -v /:/host {image}")
-        return _cmd.format(runtime=self.container_runtime,
-                           name=self.sos_container_name,
-                           image=self.container_image)
-
-    def set_cleanup_cmd(self):
-        return 'podman rm --force %s' % self.sos_container_name
-
-
-class CentOsAtomicPolicy(RedHatAtomicPolicy):
-    distro = "CentOS Atomic Host"
-    vendor = "CentOS"
-    vendor_url = "https://www.centos.org/"
-
-
-class FedoraPolicy(RedHatPolicy):
-
-    distro = "Fedora"
-    vendor = "the Fedora Project"
-    vendor_url = "https://fedoraproject.org/"
-
-    def __init__(self, sysroot=None, init=None, probe_runtime=True,
-                 remote_exec=None):
-        super(FedoraPolicy, self).__init__(sysroot=sysroot, init=init,
-                                           probe_runtime=probe_runtime,
-                                           remote_exec=remote_exec)
-
-    @classmethod
-    def check(cls, remote=''):
-        """This method checks to see if we are running on Fedora. It returns
-        True or False."""
-
-        if remote:
-            return cls.distro in remote
-
-        return os.path.isfile('/etc/fedora-release')
-
-    def fedora_version(self):
-        pkg = self.pkg_by_name("fedora-release") or \
-            self.all_pkgs_by_name_regex("fedora-release-.*")[-1]
-        return int(pkg["version"])
 
 # vim: set et ts=4 sw=4 :
