+++ sos/policies/rhel_coreos.py
@@ -1,0 +1,58 @@
+# Copyright (C) Steve Conklin <sconklin@redhat.com>
+
+# This file is part of the sos project: https://github.com/sosreport/sos
+#
+# This copyrighted material is made available to anyone wishing to use,
+# modify, copy, or redistribute it subject to the terms and conditions of
+# version 2 of the GNU General Public License.
+#
+# See the LICENSE file in the source distribution for further information.
+
+
+from sos.policies import RHELPolicy
+from sos import _sos as _
+
+
+class RedHatCoreOSPolicy(RHELPolicy):
+    distro = "Red Hat CoreOS"
+    msg = _("""\
+This command will collect diagnostic and configuration \
+information from this %(distro)s system.
+
+An archive containing the collected information will be \
+generated in %(tmpdir)s and may be provided to a %(vendor)s \
+support representative.
+""" + disclaimer_text + "%(vendor_text)s\n")
+
+    containerized = True
+    container_runtime = 'podman'
+    container_image = 'registry.redhat.io/rhel8/support-tools'
+    sos_path_strip = '/host'
+    container_version_command = 'rpm -q sos'
+
+    def __init__(self, sysroot=None, init=None, probe_runtime=True,
+                 remote_exec=None):
+        super(RedHatCoreOSPolicy, self).__init__(sysroot=sysroot, init=init,
+                                                 probe_runtime=probe_runtime,
+                                                 remote_exec=remote_exec)
+
+    def probe_preset(self):
+        # As of the creation of this policy, RHCOS is only available for
+        # RH OCP environments.
+        return self.find_preset(RHOCP)
+
+    def create_sos_container(self):
+        _cmd = ("{runtime} run -di --name {name} --privileged --ipc=host"
+                " --net=host --pid=host -e HOST=/host -e NAME={name} -e "
+                "IMAGE={image} -v /run:/run -v /var/log:/var/log -v "
+                "/etc/machine-id:/etc/machine-id -v "
+                "/etc/localtime:/etc/localtime -v /:/host {image}")
+        return _cmd.format(runtime=self.container_runtime,
+                           name=self.sos_container_name,
+                           image=self.container_image)
+
+    def set_cleanup_cmd(self):
+        return 'podman rm --force %s' % self.sos_container_name
+
+
+# vim: set et ts=4 sw=4 :
